chore: show error when user tries to delete customer with invoices

+++ UTB-AP5PW-Invoicer.Application/Features/Customers/Commands/Delete/DeleteCustomerCommandHandler.cs
@@ -1,4 +1,5 @@
 ﻿using MediatR;
+using Microsoft.EntityFrameworkCore;
 using UTB_AP5PW_Invoicer.Infrastructure.Data;
 
 namespace UTB_AP5PW_Invoicer.Application.Features.Customers.Commands.Delete
@@ -10,6 +11,10 @@ namespace UTB_AP5PW_Invoicer.Application.Features.Customers.Commands.Delete
         {
             var customer = await dbContext.Customers.FindAsync([request.Id], cancellationToken);
             if (customer == null) return false;
+
+            var invoiceCount = await dbContext.Invoices
+                .CountAsync(i => i.CustomerId == request.Id, cancellationToken);
+            if (invoiceCount > 0) return false;
 
             dbContext.Customers.Remove(customer);
             await dbContext.SaveChangesAsync(cancellationToken);
